n-ary-tree-preorder-traversal/main_test.go
package n_ary_tree_preorder_traversal

import (
	"reflect"
	"testing"
)

func Test_preorder(t *testing.T) {
	type args struct {
		root *Node
	}
	tests := []struct {
		name string
		args args
		want []int
	}{
		{
			name: "equal",
			args: args{
				root: createNAryTree(),
			},
			want: []int{1, 3, 5, 6, 2, 4},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			// if got := preorder(tt.args.root); !reflect.DeepEqual(got, tt.want) {
			// 	t.Errorf("preorder() = %v, want %v", got, tt.want)
			// }

			if got := preorderS2(tt.args.root); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("preorder() = %v, want %v", got, tt.want)
			}
		})
	}
}

func createNAryTree() *Node {
	return &Node{
		Val: 1,
		Children: []*Node{
			&Node{
				Val: 3,
				Children: []*Node{
					&Node{
						Val: 5,
					},
					&Node{
						Val: 6,
					},
				},
			},
			&Node{
				Val: 2,
			},
			&Node{
				Val: 4,
			},
		},
	}
}
